+++ changelog.md
@@ -3,6 +3,9 @@ Liferay Plugin for IntelliJ
 
 All notable changes to this project will be documented in this file.
 
+## 0.0.18
+- Feature: add freemarker taglib deprecation inspections
+
 ## 0.0.17
 - Feature: Code completion features for client-extensions.yaml
 - Feature: Code completion features for rest-config.yaml